+++ packages/api/src/modules/users/resolvers/mutation.complete-registration.ts
@@ -1,10 +1,9 @@
 import { builder, createLdapUser, prisma, yearTier } from '#lib';
 import { DateTimeScalar } from '#modules/global';
 import { NotificationChannel, type Major, type UserCandidate } from '@centraverse/db/prisma';
-import { hash } from 'argon2';
 import { ZodError } from 'zod';
 import { notify } from '../../notifications/utils/send.js';
-import { completeRegistration } from '../index.js';
+import { completeRegistration, hashPassword } from '../index.js';
 // TODO rename registration to signup
 
 builder.mutationField('completeRegistration', (t) =>
@@ -49,11 +48,9 @@ builder.mutationField('completeRegistration', (t) =>
         data: {
           action: 'complete',
           area: 'signups',
-
           message: `Complétion de l'inscription de ${firstName} ${lastName} ${yearTier(
             graduationYear,
           ).toString()}A ${phone}`,
-
           target: `token ${token}`,
         },
       });
@@ -69,11 +66,14 @@ builder.mutationField('completeRegistration', (t) =>
             graduationYear,
             lastName,
             phone,
-            password: await hash(password),
+            password: await hashPassword(password),
             cededImageRightsToTVn7,
             apprentice,
           },
-          include: { major: true },
+          include: {
+            major: { include: { schools: true } },
+            usingQuickSignup: { include: { school: { include: { majors: true } } } },
+          },
         }),
       );
       const userOrCandidate: (typeof user | UserCandidate) & { major?: Major | undefined | null } =
@@ -85,7 +85,30 @@ builder.mutationField('completeRegistration', (t) =>
 
       const needsVerification = !user;
 
-      await notify(await prisma.user.findMany({ where: { admin: true } }), {
+      const adminsResponsibleForThisSignup = await prisma.user.findMany({
+        where: {
+          OR: [
+            { admin: true },
+            ...(userOrCandidate.majorId
+              ? [
+                  {
+                    adminOfStudentAssociations: {
+                      some: {
+                        school: {
+                          majors: {
+                            some: { id: userOrCandidate.majorId },
+                          },
+                        },
+                      },
+                    },
+                  },
+                ]
+              : []),
+          ],
+        },
+      });
+
+      await notify(adminsResponsibleForThisSignup, {
         title: needsVerification ? `Inscription en attente de validation` : `Nouvelle inscription!`,
         body:
           `${userOrCandidate.email} (${userOrCandidate.firstName} ${userOrCandidate.lastName}, ${
@@ -94,7 +117,7 @@ builder.mutationField('completeRegistration', (t) =>
           (needsVerification ? `a fait une demande d'inscription` : `s'est inscrit·e!`),
         data: {
           channel: NotificationChannel.Other,
-          goto: needsVerification ? '/signups' : `/@${user.uid}`,
+          goto: needsVerification ? `/signups/edit/${userOrCandidate.email}` : `/@${user.uid}`,
           group: undefined,
         },
       });