+++ packages/api/src/modules/users/resolvers/mutation.use-password-reset.ts
@@ -1,7 +1,7 @@
 import { TYPENAMES_TO_ID_PREFIXES, builder, prisma, resetLdapUserPassword } from '#lib';
+import { hashPassword } from '#modules/users/utils';
 
 import { CredentialType as PrismaCredentialType } from '@centraverse/db/prisma';
-import { hash } from 'argon2';
 // TODO maybe rename to reset-password
 // OR merge with mutation.change-password
 
@@ -44,7 +44,7 @@ builder.mutationField('usePasswordReset', (t) =>
             },
             create: {
               type: PrismaCredentialType.Password,
-              value: await hash(newPassword),
+              value: await hashPassword(newPassword),
             },
           },
         },