+++ packages/api/src/modules/users/utils/index.ts
@@ -4,7 +4,9 @@
 
 export * from './names.js';
 export * from './password-resets.js';
+export * from './passwords.js';
 export * from './permissions.js';
 export * from './quick-signup.js';
 export * from './registration.js';
+export * from './school-emails.js';
 export * from './uid.js';