+++ packages/app/src/routes/(app)/login/+page.svelte
@@ -41,6 +41,7 @@
           {
             '__typename': true,
             '...on Error': { message: true },
+            '...on AwaitingValidationError': { __typename: true },
             '...on MutationLoginSuccess': {
               data: { token: true, expiresAt: true, user: sessionUserQuery() },
             },
@@ -50,6 +51,13 @@
 
       if (login.__typename === 'Error') {
         errorMessages = [login.message];
+        return;
+      }
+
+      if (login.__typename === 'AwaitingValidationError') {
+        errorMessages = [
+          "Ton compte n'a pas encore été validé par l'équipe d'administration de ton AE. Encore un peu de patience 😉",
+        ];
         return;
       }