+++ packages/app/src/routes/(app)/register/continue/+page.svelte
@@ -91,6 +91,7 @@
             {
               '__typename': true,
               '...on Error': { message: true },
+              '...on AwaitingValidationError': { message: true },
               '...on MutationLoginSuccess': {
                 data: { token: true, expiresAt: true, user: sessionUserQuery() },
               },
@@ -121,17 +122,17 @@
   <form title="Finaliser mon inscription" on:submit|preventDefault={register}>
     {#if data.userCandidate.emailValidated}
       <Alert theme="success" inline>
-        <strong>Votre inscription est en attente de validation manuelle.</strong><br />
-        Cependant, vous pouvez toujours compléter ou corriger les informations ci-dessous.
+        <strong>Ton inscription est en attente de validation par ton AE.</strong><br />
+        Tu peux toujours corriger des informations en attendant.
       </Alert>
-    {:else if isStudent && data.userCandidate.schoolUid === null && !data.userCandidate.usingQuickSignup}
+    {:else if isStudent && data.userCandidate.needsManualValidation}
       <Alert theme="warning" inline>
         <strong>
-          Votre compte n'est pas encore lié à une école, votre inscription sera validée
-          manuellement.
+          Tu n'a pas renseigné ton adresse e-mail universitaire. Une personne de l'équipe
+          d'administration de ton AE devra valider manuellement ton inscription.
         </strong>
         <br />
-        Si vous possédez une adresse universitaire, vous pouvez
+        Si tu as accès à toi boîte mail universitaire, tu peux
         <a href="..">recommencer l'inscription</a> avec celle-ci.
       </Alert>
     {/if}
@@ -220,7 +221,9 @@
     </div>
     <InputCheckbox bind:value={cededImageRightsToTVn7} label="Je cède mon droit à l'image à TVn7" />
     <p class="typo-details">
-      Cela revient à remplir et signer <a href="/cessation-droit-image-tvn7.pdf">ce document</a>
+      Cela revient à remplir et signer <a target="_blank" href="/cessation-droit-image-tvn7.pdf"
+        >ce document</a
+      >
     </p>
     <section class="optional-info">
       <h2>Informations Personnelles</h2>
@@ -254,8 +257,8 @@
   <Alert theme="success">
     <h3>Demande enregistrée&nbsp;!</h3>
     <p>
-      Votre inscription sera validée manuellement et vous recevrez un email une fois votre compte
-      validé.
+      Tu recevra un email une fois que ton compte sera validé par l'équipe d'administration de ton
+      AE.
     </p>
     <p><a href="/">Retourner à l'accueil.</a></p>
   </Alert>