+++ packages/app/src/routes/(app)/signups/+page.svelte
@@ -7,6 +7,7 @@
   import ButtonSecondary from '$lib/components/ButtonSecondary.svelte';
   import { tooltip } from '$lib/tooltip';
   import { toasts } from '$lib/toasts';
+  import { removeIdPrefix } from '$lib/typenames';
 
   export let data: PageData;
 
@@ -26,18 +27,34 @@
 
     try {
       loadingRegistrations.push(email);
-      let result = false;
+      let ok = false;
       if (accept) {
-        ({ acceptRegistration: result } = await $zeus.mutate({
-          acceptRegistration: [{ email }, true],
-        }));
+        const { acceptRegistration: result } = await $zeus.mutate({
+          acceptRegistration: [
+            { email },
+            {
+              '__typename': true,
+              '...on Error': { message: true },
+              '...on MutationAcceptRegistrationSuccess': {
+                data: {
+                  email: true,
+                },
+              },
+            },
+          ],
+        });
+
+        if (result.__typename === 'Error')
+          toasts.error("Erreur lors de l'acceptation de l'inscription", result.message);
+        else ok = true;
       } else {
-        ({ refuseRegistration: result } = await $zeus.mutate({
+        const { refuseRegistration: result } = await $zeus.mutate({
           refuseRegistration: [{ email, reason: why }, true],
-        }));
+        });
+        ok = result;
       }
 
-      if (result) removeRow(email);
+      if (ok) removeRow(email);
     } catch (error) {
       toasts.error("Erreur lors de la décision de l'inscription", error?.toString() ?? '');
     } finally {
@@ -48,8 +65,8 @@
 
 <h1>Demandes d'inscription</h1>
 <ul class="nobullet registrations">
-  {#each userCandidates as { email, fullName, major, graduationYear }}
-    <li>
+  {#each userCandidates as { email, fullName, major, graduationYear, id }}
+    <li id={removeIdPrefix('UserCandidate', id)}>
       <strong>{fullName}</strong>
       <span
         >{email} · {#if major}<abbr title="" use:tooltip={major.name}>{major.shortName}</abbr