+++ packages/db/seed/index.ts
@@ -1,5 +1,4 @@
 import { fakerFR } from '@faker-js/faker';
-import { hash } from 'argon2';
 import { format } from 'date-fns';
 import dichotomid from 'dichotomid';
 import { existsSync, readdirSync, statSync } from 'node:fs';
@@ -19,6 +18,9 @@ import {
   Visibility,
   type Prisma,
 } from '../src/client/default.js';
+
+// XXX: will break if we change the hashing algorithm (abstracted away in api package but not here)
+import { hash } from 'argon2';
 
 const prisma = new PrismaClient();
 
@@ -136,34 +138,44 @@ const createUid = async ({ firstName, lastName }: { firstName: string; lastName:
   return `${base}${n > 1 ? n : ''}`;
 };
 
-const schoolsData = [
+const schoolsData: Prisma.SchoolCreateInput[] = [
   {
     name: 'EAU',
     uid: 'o',
     color: '#00ffff',
     description: 'École de l’Eau',
-    address: '2 rue Charles Camichel, 31000 Toulouse', //generation par faker possible ???
+    address: faker.location.streetAddress(), //generation par faker possible ???
   },
   {
     name: 'FEU',
     uid: 'feu',
     color: '#b22222',
     description: 'École de Feu',
-    address: '2 rue Charles Camichel, 31000 Toulouse',
+    address: faker.location.streetAddress(),
+    studentMailDomain: 'etu.inp-n7.fr',
+    aliasMailDomains: ['etu.toulouse-inp.fr'],
   },
   {
     name: 'TERRE',
     uid: '3',
     color: '#5e3f13',
     description: 'École de Terre',
-    address: '2 rue Charles Camichel, 31000 Toulouse',
+    address: faker.location.streetAddress(),
+    studentMailDomain: faker.internet.domainName(),
+    aliasMailDomains: Array.from({ length: faker.number.int({ min: 0, max: 3 }) }).map(
+      faker.internet.domainName,
+    ),
   },
   {
     name: 'AIR',
     uid: 'air',
     color: '#d9eaff',
     description: 'École de l’Air',
-    address: '2 rue Charles Camichel, 31000 Toulouse',
+    address: faker.location.streetAddress(),
+    studentMailDomain: faker.internet.domainName(),
+    aliasMailDomains: Array.from({ length: faker.number.int({ min: 0, max: 3 }) }).map(
+      faker.internet.domainName,
+    ),
   },
 ];